% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{final_revision_df}
\alias{final_revision_df}
\title{Combine integration and clustering data}
\usage{
final_revision_df(integration_path, clustering_path)
}
\arguments{
\item{integration_path}{Path to CSV file with columns: cell, integration_cluster}

\item{clustering_path}{Path to CSV file with columns: cell, cell_type, sample_type}
}
\value{
none
}
\description{
Reads and merges integration and clustering tables by cell ID, ensuring consistent ID formatting.
}
