% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_functions.R
\name{final_revision_dt}
\alias{final_revision_dt}
\title{Combine integration and clustering data}
\usage{
final_revision_dt(integration_path, clustering_path)
}
\arguments{
\item{integration_path}{Path to CSV file with columns: cell, integration_cluster}

\item{clustering_path}{Path to CSV file with columns: cell, cell_type, sample_type}
}
\value{
files created in Output/
}
\description{
provide a new file where cell type, cells and integration clusters are combined,
provide a file where the number of each cell type is indicated for each cluster,
provide summary table showing cell types associated to integration clusters and also their
association to normal and tumor tissue.
Provide a plot describing the distribution of the cell type in normal/tumor tissue given the
integration clusters.
Provide a normalized % for the cell types in each of the integration clusters for normal and
tumor specimen.
}
